Release 7.3.1 (#14)
* Update project to latest tag saceriam-jboss-7.3.1

* Update SNAPSHOT version

* [maven-release-plugin] prepare release saceriam-jboss-7.3.1

* [maven-release-plugin] prepare for next development iteration

* Update pom.xml

---------

+++ SacerIam-ejb/src/main/java/it/eng/saceriam/ws/replicaOrganizzazione/dto/TipoDato.java
@@ -19,7 +19,7 @@ package it.eng.saceriam.ws.replicaOrganizzazione.dto;
  */
 public class TipoDato {
     private String nmClasseTipoDato;
-    private Integer idTipoDatoApplic;
+    private Long idTipoDatoApplic;
     private String nmTipoDato;
     private String dsTipoDato;
 
@@ -31,11 +31,11 @@ public class TipoDato {
 	this.nmClasseTipoDato = nmClasseTipoDato;
     }
 
-    public Integer getIdTipoDatoApplic() {
+    public Long getIdTipoDatoApplic() {
 	return idTipoDatoApplic;
     }
 
-    public void setIdTipoDatoApplic(Integer idTipoDatoApplic) {
+    public void setIdTipoDatoApplic(Long idTipoDatoApplic) {
 	this.idTipoDatoApplic = idTipoDatoApplic;
     }
